+++ packages/kevin_flutter_accounts/kevin_flutter_accounts_android/android/src/main/kotlin/eu/kevin/flutter/accounts/entity/AccountSessionConfigurationEntity.kt
@@ -11,5 +11,4 @@ internal data class AccountSessionConfigurationEntity(
     val bankFilter: List<String>,
     val preselectedBank: String?,
     val skipBankSelection: Boolean,
-    val accountLinkingType: String,
 )