+++ packages/kevin_flutter_accounts/kevin_flutter_accounts_ios/ios/Classes/Entity/AccountSessionConfigurationEntity.swift
@@ -8,5 +8,4 @@ internal struct AccountSessionConfigurationEntity: Codable {
     let bankFilter: [String]
     let preselectedBank: String?
     let skipBankSelection: Bool
-    let accountLinkingType: String
 }