+++ packages/kevin_flutter_in_app_payments/kevin_flutter_in_app_payments_platform_interface/lib/src/model/kevin_payments_models.dart
@@ -1,4 +1,3 @@
 export 'kevin_session_result_payment_success.dart';
 export 'payment/kevin_payment_session_configuration.dart';
-export 'payment/kevin_payment_type.dart';
 export 'payment/kevin_payments_configuration.dart';