+++ packages/kevin_flutter_in_app_payments/kevin_flutter_in_app_payments_platform_interface/lib/src/model/payment/kevin_payment_session_configuration.dart
@@ -1,9 +1,7 @@
 import 'package:kevin_flutter_core/kevin_flutter_core.dart';
-import 'package:kevin_flutter_in_app_payments_platform_interface/src/model/payment/kevin_payment_type.dart';
 
 class KevinPaymentSessionConfiguration {
   final String paymentId;
-  final KevinPaymentType paymentType;
   final KevinCountry? preselectedCountry;
   final bool disableCountrySelection;
   final List<KevinCountry> countryFilter;
@@ -14,7 +12,6 @@ class KevinPaymentSessionConfiguration {
 
   const KevinPaymentSessionConfiguration({
     required this.paymentId,
-    this.paymentType = KevinPaymentType.bank,
     this.preselectedCountry,
     this.disableCountrySelection = false,
     this.countryFilter = const [],
@@ -27,7 +24,6 @@ class KevinPaymentSessionConfiguration {
   Map<String, dynamic> toMap() {
     return {
       'paymentId': paymentId,
-      'paymentType': paymentType.name,
       'preselectedCountry': preselectedCountry?.iso,
       'disableCountrySelection': disableCountrySelection,
       'countryFilter': countryFilter.map((c) => c.iso).toList(),