+++ packages/kevin_flutter_in_app_payments/kevin_flutter_in_app_payments_platform_interface/lib/src/model/payment/kevin_payment_type.dart
@@ -1,1 +1,0 @@
-enum KevinPaymentType { bank, card }