+++ packages/kevin_flutter_accounts/kevin_flutter_accounts_android/CHANGELOG.md
@@ -1,3 +1,12 @@
+## 1.2.0
+
+* **BREAKING**: KevinAccountLinkingType was deprecated and removed, only bank linking is accepted
+* Upgrade:
+    - `eu.kevin.android:core:2.7.1`
+    - `eu.kevin.android:accounts:2.7.1`
+    - `kevin_flutter_core: 1.2.0`
+    - `kevin_flutter_accounts_platform_interface: 1.2.0`
+
 ## 1.1.0
 
 * Update Dart min version to 3.0.0